[FIX] Version numbering
The previous version numbering schem was wrong.
The patch number was increased when feat was added
and the minor number was unused.
Now the minor number is set to 3 (the actual number of FEAT
commit on the beta branch) and the patch version is set to
one (cosidering this commit as a bugfix).